consolidate NetteTester to PHPUnit tests to one
