+++ packages/NetteTesterToPHPUnit/tests/Rector/Class_/NetteTesterClassToPHPUnitClassRector/Fixture/assert_true.php.inc
