+++ packages/NetteTesterToPHPUnit/tests/Rector/Class_/NetteTesterClassToPHPUnitClassRector/Fixture/assert_type.php.inc
@@ -36,7 +36,7 @@ namespace Rector\NetteTesterToPHPUnit\Tests\Rector\StaticCall\NetteAssertToPHPUn
 use Rector\NetteTesterToPHPUnit\Tests\Rector\Class_\NetteTesterClassToPHPUnitClassRector\Source\NetteTesterTestCase;
 use Tester\Assert;
 
-class AssertTypeTest extends NetteTesterTestCase
+class AssertTypeTest extends \PHPUnit\Framework\TestCase
 {
     public function testFunctionality()
     {
